cast numbers into int form for flat JSON output, fixes #98
When interpreting JSON the way it is done here, type information
is not used, but JSON numbers are cast into float64. This results
in strange output format for large numbers, such as times (in the
Unix timestamp format) being displayed with base+exponential
representation (e.g. "1.427376789671e+12")

The API interface does not actually have any float64 numbers,
only int/int64, thus, at least at this point, it is safe to cast
everly float64 found into int by using a 0-precision representation.

On the long term, a better solution would be to unmarshal the API
response using the LabsResponse struct, and iterate over that
to create a flattened JSON output. This would needs more refactoring,
but would correctly infer the types of every field.