new feature : creates txt file for all hostname with their json string in it
--output-host is a new boolean switch which will allow the user to create txt files based on the hostname containing the API result provided by the API of ssllabs.

Files will be created in the same folder where the code resides. This feature is helpful for multiple hostnames. E.g If www.ssllabs.com is scanned with --output-host switch then a file "www.ssllabs.com.txt" will be created
Note: This will rewrite the old file with same name